feat: CreateCommentRequest 확장 추가

+++ src/main/kotlin/ga/injuk/commentor/domain/model/CommentPart.kt
@@ -6,6 +6,6 @@ data class CommentPart(
     val content: List<Content>,
 ) {
     data class Attrs(
-        val level: Int?,
+        val level: Long?,
     )
 }
